refactor(pma-routes): extract managed thread scope service and reorganize PMA route layer
- Extract managed_thread_scope.py (646 lines) from monolithic route helpers
- Move automation_adapter.py to surfaces/web/services/pma/automation.py
- Harden managed thread followup and tail stream with scope context
- Update test helpers for new service imports
