+++ src/codex_autorunner/surfaces/web/routes/pma_routes/__init__.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
+from ...services.pma.automation import notify_managed_thread_terminal_transition
 from .action_manifest import build_action_manifest_routes
-from .automation_adapter import notify_managed_thread_terminal_transition
 from .chat_events import build_chat_event_routes
 from .chat_runtime import build_chat_runtime_router
 from .history_files_docs import build_history_files_docs_router
